chore: Move client into package and add NewPoolFactory function
